Move files into package and add placeholder unit tests
