+++ src/mqm/geo_process.py
@@ -3,7 +3,7 @@ import numpy as np
 import sys
 import os
 from area import area
-from utility import Utility
+from .utility import Utility
 import argparse
