+++ tests/mqm/test_geo_process.py
@@ -1,0 +1,7 @@
+import mqm
+
+
+def test_initialize():
+    initialized = mqm.GeoProcessor('')
+    assert initialized is not None
+
