+++ tests/mqm/test_kd_tree.py
@@ -1,0 +1,6 @@
+import mqm
+
+
+def test_init():
+    initialized = mqm.kdTree(None, None, None, None)
+    assert initialized is not None
