+++ tests/mqm/test_mqm_tool.py
@@ -1,0 +1,5 @@
+import mqm
+
+
+def test_import():
+    assert mqm.road_count is not None
