+++ tests/mqm/test_utility.py
@@ -1,0 +1,6 @@
+import mqm
+
+
+def test_initialize():
+    initialized = mqm.Utility()
+    assert initialized is not None
