+++ tox.ini
@@ -3,4 +3,7 @@ envlist = py37
 
 [testenv]
 deps = pytest
+    matplotlib
+    area
+    numpy
 commands = pytest
